Add checks to avoid re-dl-ing

+++ src/data/downloader.py
@@ -37,12 +37,13 @@ def download(post_id, post_url, output_dir=OUTPUT_DIR):
     *_, extension = os.path.basename(urlparse(post_url).path).split('.')
     output_filename = os.path.join(output_dir, f"{post_id}.{extension}")
     
-    # Download the file from `url` and save it locally under `output_filename`: 
-    try:   
-        with urlopen(post_url) as response, open(output_filename, 'wb') as out_file:
-            shutil.copyfileobj(response, out_file)
-    except:
-        pass
+    # Download the file from `url` and save it locally under `output_filename`
+    if not os.path.isfile(output_filename):
+        try:   
+            with urlopen(post_url) as response, open(output_filename, 'wb') as out_file:
+                shutil.copyfileobj(response, out_file)
+        except:
+            pass
     
     print(f"Downloaded {output_filename}")
